Create 12-http.js file
