--- app.js
@@ -1,17 +1,0 @@
-const http = require('http');
-
-const server = http.createServer((req, res) => {
-  if (req.url === '/') {
-    res.end('Welcome to our home page');
-  } else if (req.url === '/about') {
-    res.end('Here is our short history');
-  } else {
-    res.end(`
-      <h1>oops!</h1>
-      <p>We can't seem to find the page you are looking for</p>
-      <a href="/">Back Home</a>
-    `);
-  }
-});
-
-server.listen(5000);
